requirements.txt
ago==0.0.93
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
awscli==1.16.201
beautifulsoup4==4.7.1
botocore==1.12.191
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
docutils==0.14
dotmap==1.3.8
elasticsearch==7.0.2
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-SocketIO==4.1.0
gunicorn==19.9.0
hjson==3.0.1
hurry.filesize==0.9
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
langdetect==1.0.7
lxml==4.3.4
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.4
numpy==1.16.4
oauthlib==3.0.2
parsel==1.5.1
Pillow==6.1.0
plac==1.0.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.8.2.post1
python-socketio==4.2.0
PyYAML==5.1
queuelib==1.5.0
readability-lxml==0.7.1
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
Scrapy==1.7.1
service-identity==18.1.0
six==1.12.0
sklearn==0.0
socketIO-client==0.7.2
soupsieve==1.9.2
stop-words==2018.7.23
textblob==0.15.3
timeloop==1.0.2
tinysegmenter==0.3
tldextract==2.2.1
tweepy==3.8.0
Twisted==19.2.1
Unidecode==1.1.1
urllib3==1.25.3
w3lib==1.20.0
warcio==1.7.1
websocket-client==0.56.0
Werkzeug==0.15.5
zope.interface==4.6.0
